Refactor spec initialization
