+++ spec/spec_helper.rb
@@ -97,4 +97,12 @@ RSpec.configure do |config|
   #   # test failures related to randomization by passing the same `--seed` value
   #   # as the one that triggered the failure.
   #   Kernel.srand config.seed
+
+  config.before(:suite) do
+    SpecUtils::SparkleSetup.initial_install
+  end
+
+  config.after(:suite) do
+    SpecUtils::SparkleSetup.final_deinstall
+  end
 end
